Fix disk migration and fix sine dispatcher plugins dependencies.
Remove entire list of dependencies on other dispatcher plugins.
This is ok and harmless because of the following two reasons:
1. All migration work is done via tasks and tasks do not run till all dispatcher
plugins are loaded.
2. The MasterMigrateTask task is run via an event handler on serviced's
ready event so by that time all plugins will be loaded and up anyways.

While here fix some dispatcher plugin dependencies that I found to be incorrect or excessive.

Ticket: #19568